Refactoring of the loan module.
Refactoring of the loan module.
