--- frontend/pages/finance/prestamos.vue
@@ -1,9 +1,0 @@
-import { Deduccion } from '../.nuxt/components';
-<template>
-    <NuxtLayout>
-            <Deduccion />
-    </NuxtLayout>
-</template>
-
-<script>
-</script>
